+++ src/ContaoManager/Plugin.php
@@ -1,12 +1,15 @@
 <?php
-/**
+
+declare(strict_types=1);
+
+/*
  * theme components bundle for Contao Open Source CMS
  *
- * Copyright (C) 2017 pdir / digital agentur <develop@pdir.de>
+ * Copyright (C) 2022 pdir / digital agentur <develop@pdir.de>
  *
  * @package    contao-themes-net/theme-components-bundle
  * @link       https://github.com/contao-themes-net/theme-components-bundle
- * @license    pdir contao theme licence
+ * @license    LGPL-3.0+
  * @author     pdir GmbH <develop@pdir.de>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -16,11 +19,11 @@
 namespace ContaoThemesNet\ThemeComponentsBundle\ContaoManager;
 
 use Contao\CoreBundle\ContaoCoreBundle;
-use ContaoThemesNet\ThemeComponentsBundle\ThemeComponentsBundle;
-use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
+use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 use ContaoDD\AdvancedClassesBundle\ContaoDDAdvancedClassesBundle;
+use ContaoThemesNet\ThemeComponentsBundle\ThemeComponentsBundle;
 
 /**
  * Plugin for the Contao Manager.
@@ -38,7 +41,7 @@ class Plugin implements BundlePluginInterface
             BundleConfig::create(ThemeComponentsBundle::class)
                 ->setLoadAfter([
                     ContaoCoreBundle::class,
-                    ContaoDDAdvancedClassesBundle::class
+                    ContaoDDAdvancedClassesBundle::class,
                 ]),
         ];
     }