+++ src/Element/FeatureElement.php
@@ -1,21 +1,38 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * theme components bundle for Contao Open Source CMS
+ *
+ * Copyright (C) 2022 pdir / digital agentur <develop@pdir.de>
+ *
+ * @package    contao-themes-net/theme-components-bundle
+ * @link       https://github.com/contao-themes-net/theme-components-bundle
+ * @license    LGPL-3.0+
+ * @author     pdir GmbH <develop@pdir.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace ContaoThemesNet\ThemeComponentsBundle\Element;
 
 class FeatureElement extends \ContentElement
 {
     /**
-     * Template
+     * Template.
+     *
      * @var string
      */
     protected $strTemplate = 'ce_cthemes_feature_element';
 
     /**
-     * Generate the content element
+     * Generate the content element.
      */
-    protected function compile()
+    protected function compile(): void
     {
-        if($this->ct_featureElement_customTpl != "") {
+        if ('' !== $this->ct_featureElement_customTpl) {
             $this->Template->setName($this->ct_featureElement_customTpl);
         }
 
@@ -25,8 +42,8 @@ class FeatureElement extends \ContentElement
         // overwrite link target
         $this->Template->target = '';
         $this->Template->rel = '';
-        if ($this->target)
-        {
+
+        if ($this->target) {
             $this->Template->target = ' target="_blank"';
             $this->Template->rel = ' rel="noreferrer noopener"';
         }