+++ src/Element/WrapperStartElement.php
@@ -1,28 +1,48 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * theme components bundle for Contao Open Source CMS
+ *
+ * Copyright (C) 2022 pdir / digital agentur <develop@pdir.de>
+ *
+ * @package    contao-themes-net/theme-components-bundle
+ * @link       https://github.com/contao-themes-net/theme-components-bundle
+ * @license    LGPL-3.0+
+ * @author     pdir GmbH <develop@pdir.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace ContaoThemesNet\ThemeComponentsBundle\Element;
+
+use Contao\BackendTemplate;
 
 class WrapperStartElement extends \ContentElement
 {
     /**
-     * Template
+     * Template.
+     *
      * @var string
      */
     protected $strTemplate = 'ce_cthemes_wrapper_start';
 
-
     /**
-     * Generate the content element
+     * Generate the content element.
      */
-    protected function compile()
+    protected function compile(): void
     {
-        if (TL_MODE == 'BE')
-        {
+        if (TL_MODE === 'BE') {
             $this->strTemplate = 'be_wildcard';
             /** @var BackendTemplate|object $objTemplate */
-            $objTemplate = new \BackendTemplate($this->strTemplate);
+            $objTemplate = new BackendTemplate($this->strTemplate);
             $this->Template = $objTemplate;
-            if($this->ct_wrapper_name != "") $this->Template->wildcard = "### " . $this->ct_wrapper_name. " ###";
+
+            if ('' !== $this->ct_wrapper_name) {
+                $this->Template->wildcard = '### '.$this->ct_wrapper_name.' ###';
+            }
         }
     }
 }