+++ src/Resources/contao/config/config.php
@@ -1,5 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * theme components bundle for Contao Open Source CMS
+ *
+ * Copyright (C) 2022 pdir / digital agentur <develop@pdir.de>
+ *
+ * @package    contao-themes-net/theme-components-bundle
+ * @link       https://github.com/contao-themes-net/theme-components-bundle
+ * @license    LGPL-3.0+
+ * @author     pdir GmbH <develop@pdir.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 use ContaoThemesNet\ThemeComponentsBundle\Element\ContentBoxElement;
 use ContaoThemesNet\ThemeComponentsBundle\Element\FeatureElement;
 use ContaoThemesNet\ThemeComponentsBundle\Element\PriceBoxElement;
@@ -10,9 +26,9 @@ use ContaoThemesNet\ThemeComponentsBundle\Element\WrapperStopElement;
 use ContaoThemesNet\ThemeComponentsBundle\Module\ModalModule;
 
 // Insert the mate theme category
-array_insert($GLOBALS['TL_CTE'], 1, array('contaoThemesNet' => array()));
+array_insert($GLOBALS['TL_CTE'], 1, ['contaoThemesNet' => []]);
 
-/**
+/*
  * Add content element
  */
 $GLOBALS['TL_CTE']['contaoThemesNet']['ct_contentBox'] = ContentBoxElement::class;
@@ -23,13 +39,13 @@ $GLOBALS['TL_CTE']['contaoThemesNet']['ct_teaserBox'] = TeaserBoxElement::class;
 $GLOBALS['TL_CTE']['contaoThemesNet']['ct_wrapperStart'] = WrapperStartElement::class;
 $GLOBALS['TL_CTE']['contaoThemesNet']['ct_wrapperStop'] = WrapperStopElement::class;
 
-/**
+/*
  * Add modules
  */
 
 $GLOBALS['FE_MOD']['contaoThemesNet']['ct_modal'] = ModalModule::class;
 
-/**
+/*
  * Wrapper elements
  */
 $GLOBALS['TL_WRAPPERS']['start'][] = 'ct_wrapperStart';