+++ src/Resources/contao/languages/de/default.php
@@ -1,5 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * theme components bundle for Contao Open Source CMS
+ *
+ * Copyright (C) 2022 pdir / digital agentur <develop@pdir.de>
+ *
+ * @package    contao-themes-net/theme-components-bundle
+ * @link       https://github.com/contao-themes-net/theme-components-bundle
+ * @license    LGPL-3.0+
+ * @author     pdir GmbH <develop@pdir.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 $GLOBALS['TL_LANG']['CTE']['contaoThemesNet'] = 'Contao Themes';
 
 $GLOBALS['TL_LANG']['CTE']['ct_contentBox'][0] = 'Content Box';