+++ src/Resources/contao/languages/en/tl_module.php
@@ -1,10 +1,27 @@
 <?php
+
+declare(strict_types=1);
+
+/*
+ * theme components bundle for Contao Open Source CMS
+ *
+ * Copyright (C) 2022 pdir / digital agentur <develop@pdir.de>
+ *
+ * @package    contao-themes-net/theme-components-bundle
+ * @link       https://github.com/contao-themes-net/theme-components-bundle
+ * @license    LGPL-3.0+
+ * @author     pdir GmbH <develop@pdir.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 $GLOBALS['TL_LANG']['tl_module']['modal_link_legend'] = 'Link settings';
 $GLOBALS['TL_LANG']['tl_module']['modal_headline_legend'] = 'Modal headline';
 $GLOBALS['TL_LANG']['tl_module']['modal_text_legend'] = 'Modal content';
 
-$GLOBALS['TL_LANG']['tl_module']['modal_linkText'] = ['Link text','Here you can enter a link text.'];
-$GLOBALS['TL_LANG']['tl_module']['modal_linkClass'] = ['Link classes','Here you can add more classes to the link.'];
-$GLOBALS['TL_LANG']['tl_module']['modal_text'] = ['Text','Here you can enter the content for the modal.'];
-$GLOBALS['TL_LANG']['tl_module']['modal_customTpl'] = ['Custom template','Here you can choose a custom template.'];
-$GLOBALS['TL_LANG']['tl_module']['modal_class'] = ['CSS classes for the modal','Here you can add more classes to the modal.'];
+$GLOBALS['TL_LANG']['tl_module']['modal_linkText'] = ['Link text', 'Here you can enter a link text.'];
+$GLOBALS['TL_LANG']['tl_module']['modal_linkClass'] = ['Link classes', 'Here you can add more classes to the link.'];
+$GLOBALS['TL_LANG']['tl_module']['modal_text'] = ['Text', 'Here you can enter the content for the modal.'];
+$GLOBALS['TL_LANG']['tl_module']['modal_customTpl'] = ['Custom template', 'Here you can choose a custom template.'];
+$GLOBALS['TL_LANG']['tl_module']['modal_class'] = ['CSS classes for the modal', 'Here you can add more classes to the modal.'];