+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.3.9](https://github.com/contao-themes-net/theme-components-bundle/tree/1.3.9) – 2022-03-25
+
+- [fix] add visibility legend for content box and teaser box
+- [fix] add protected legend and guests field in dca
+
 ## [1.3.8](https://github.com/contao-themes-net/theme-components-bundle/tree/1.3.8) – 2021-05-10
 
 - [fix] remove link for image in slider element